Reverted to old ECS

+++ Amalgamation/Core/Platform/Platform.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
 using uint64 = unsigned long long int;
-using uint32 = unsigned long      int;
+using uint32 = unsigned           int;
 using uint16 = unsigned short     int;
 using uint8  = unsigned char;
 
 using int64  = signed long long int;
-using int32  = signed long      int;
+using int32  = signed           int;
 using int16  = signed short     int;
 using int8   = signed char;
